Mainly edited agent so that it's compatible with Micah's branch. Previously I had a seperate agent_micah, which contained Micah's GreedyHumanModel (GHM). Now I've put Micah's GHM into agent.py, and renamed mine GreedyHumanModel_pk. (they are both used for different things, because I've never checked if Micah's is compatible with my usage of GHM)
