+++ overcooked_ai_py/agents/agent.py
@@ -1,5 +1,4 @@
-import scipy
-import itertools
+import itertools, math
 import numpy as np
 from collections import defaultdict
 import random
@@ -9,12 +8,32 @@ from overcooked_ai_py.mdp.actions import Action, Direction
 from overcooked_ai_py.mdp.overcooked_mdp import OvercookedState
 from overcooked_ai_py.mdp.overcooked_env import OvercookedEnv
 from overcooked_ai_py.planning.planners import MediumLevelPlanner, Heuristic
+from overcooked_ai_py.planning.search import SearchTree
 
 
 class Agent(object):
 
     def action(self, state):
+        """
+        Should return an action, and an action info dictionary.
+        If collecting trajectories of the agent with OvercookedEnv, the action
+        info data will be included in the trajectory data under `ep_infos`.
+
+        This allows agents to optionally store useful information about them
+        in the trajectory for further analysis.
+        """
         return NotImplementedError()
+
+    @staticmethod
+    def a_probs_from_action(action):
+        action_idx = Action.ACTION_TO_INDEX[action]
+        return np.eye(Action.NUM_ACTIONS)[action_idx]
+
+    @staticmethod
+    def check_action_probs(action_probs, tolerance=1e-4):
+        """Check that action probabilities sum to ≈ 1.0"""
+        probs_sum = sum(action_probs)
+        assert math.isclose(probs_sum, 1.0, rel_tol=tolerance), "Action probabilities {} should sum up to approximately 1 but sum up to {}".format(list(action_probs), probs_sum)
 
     def set_agent_index(self, agent_index):
         self.agent_index = agent_index
@@ -26,9 +45,98 @@ class Agent(object):
         pass
 
 
+class AgentGroup(object):
+    """
+    AgentGroup is a group of N agents used to sample
+    joint actions in the context of an OvercookedEnv instance.
+    """
+
+    def __init__(self, *agents, allow_duplicate_agents=False):
+        self.agents = agents
+        self.n = len(self.agents)
+        for i, agent in enumerate(self.agents):
+            agent.set_agent_index(i)
+
+        if not all(a0 is not a1 for a0, a1 in itertools.combinations(agents, 2)):
+            assert allow_duplicate_agents, "All agents should be separate instances, unless allow_duplicate_agents is set to true"
+
+    def joint_action(self, state):
+        actions_and_probs_n = tuple(a.action(state) for a in self.agents)
+        return actions_and_probs_n
+
+    def set_mdp(self, mdp):
+        for a in self.agents:
+            a.set_mdp(mdp)
+
+    def reset(self):
+        for a in self.agents:
+            a.reset()
+
+
+class AgentPair(AgentGroup):
+    """
+    AgentPair is the N=2 case of AgentGroup. Unlike AgentGroup,
+    it supports having both agents being the same instance of Agent.
+
+    NOTE: Allowing duplicate agents (using the same instance of an agent
+    for both fields can lead to problems if the agents have state / history)
+    """
+
+    def __init__(self, *agents, allow_duplicate_agents=False):
+        super().__init__(*agents, allow_duplicate_agents=allow_duplicate_agents)
+        assert self.n == 2
+        self.a0, self.a1 = self.agents
+
+        if type(self.a0) is CoupledPlanningAgent and type(self.a1) is CoupledPlanningAgent:
+            print("If the two planning agents have same params, consider using CoupledPlanningPair instead to reduce computation time by a factor of 2")
+
+    def joint_action(self, state):
+        if self.a0 is self.a1:
+            # When using the same instance of an agent for self-play,
+            # reset agent index at each turn to prevent overwriting it
+            self.a0.set_agent_index(0)
+            action_and_infos_0 = self.a0.action(state)
+            self.a1.set_agent_index(1)
+            action_and_infos_1 = self.a1.action(state)
+            joint_action_and_infos = (action_and_infos_0, action_and_infos_1)
+            return joint_action_and_infos
+        else:
+            return super().joint_action(state)
+
+
+class CoupledPlanningPair(AgentPair):
+    """
+    Pair of identical coupled planning agents. Enables to search for optimal
+    action once rather than repeating computation to find action of second agent
+    """
+
+    def __init__(self, agent):
+        super().__init__(agent, agent, allow_duplicate_agents=True)
+
+    def joint_action(self, state):
+        # Reduce computation by half if both agents are coupled planning agents
+        joint_action_plan = self.a0.mlp.get_low_level_action_plan(state, self.a0.heuristic, delivery_horizon=self.a0.delivery_horizon, goal_info=True)
+
+        if len(joint_action_plan) == 0:
+            return ((Action.STAY, {}), (Action.STAY, {}))
+
+        joint_action_and_infos = [(a, {}) for a in joint_action_plan[0]]
+        return joint_action_and_infos
+
+
+
 class AgentFromPolicy(Agent):
-    
+    """
+    Defines an agent from a `state_policy` and `direct_policy` functions
+    """
+
     def __init__(self, state_policy, direct_policy, stochastic=True, action_probs=False):
+        """
+        state_policy (fn): a function that takes in an OvercookedState instance and returns corresponding actions
+        direct_policy (fn): a function that takes in a preprocessed OvercookedState instances and returns actions
+        stochastic (Bool): Whether the agent should sample from policy or take argmax
+        action_probs (Bool): Whether agent should return action probabilities or a sampled action
+        """
         self.state_policy = state_policy
         self.direct_policy = direct_policy
         self.history = []
@@ -61,16 +169,22 @@ class AgentFromPolicy(Agent):
 
 class RandomAgent(Agent):
     """
-    An agent that randomly picks actions.
-    NOTE: Does not perform interact actions
+    An agent that randomly picks motion actions.
+    NOTE: Does not perform interact actions, unless specified
     """
 
-    def __init__(self, sim_threads=None):
+    def __init__(self, sim_threads=None, interact=False):
         self.sim_threads = sim_threads
-    
+        self.interact = interact
+
     def action(self, state):
-        idx = np.random.randint(4)
-        return Action.ALL_ACTIONS[idx]
+        action_probs = np.zeros(Action.NUM_ACTIONS)
+        legal_actions = Action.MOTION_ACTIONS
+        if self.interact:
+            legal_actions.append(Action.INTERACT)
+        legal_actions_indices = np.array([Action.ACTION_TO_INDEX[motion_a] for motion_a in legal_actions])
+        action_probs[legal_actions_indices] = 1 / len(legal_actions_indices)
+        return Action.sample(action_probs), {"action_probs": action_probs}
 
     def direct_action(self, obs):
         return [np.random.randint(4) for _ in range(self.sim_threads)]
@@ -82,30 +196,40 @@ class StayAgent(Agent):
         self.sim_threads = sim_threads
     
     def action(self, state):
-        #TODO RS: This used to be Direction.STAY
-        return Action.STAY
+        a = Action.STAY
+        return a, {}
 
     def direct_action(self, obs):
         return [Action.ACTION_TO_INDEX[Action.STAY]] * self.sim_threads
 
 
 class FixedPlanAgent(Agent):
-    
+    """
+    An Agent with a fixed plan. Returns Stay actions once pre-defined plan has terminated.
+    # NOTE: Assumes that calls to action are sequential (agent has history)
+    """
+
     def __init__(self, plan):
         self.plan = plan
         self.i = 0
     
     def action(self, state):
         if self.i >= len(self.plan):
-            return None
+            return Action.STAY, {}
         curr_action = self.plan[self.i]
-
-        # NOTE: Assumes that calls to action are sequential
         self.i += 1
-        return curr_action
+        return curr_action, {}
+
+    def reset(self):
+        self.i = 0
 
 
 class CoupledPlanningAgent(Agent):
+    """
+    An agent that uses a joint planner (mlp, a MediumLevelPlanner) to find near-optimal
+    plans. At each timestep the agent re-plans under the assumption that the other agent
+    is also a CoupledPlanningAgent, and then takes the first action in the plan.
+    """
 
     def __init__(self, mlp, delivery_horizon=2, heuristic=None):
         self.mlp = mlp
@@ -120,28 +244,30 @@ class CoupledPlanningAgent(Agent):
             print("COUPLED PLANNING FAILURE")
             self.mlp.failures += 1
             return Direction.ALL_DIRECTIONS[np.random.randint(4)]
-        return joint_action_plan[0][self.agent_index] if len(joint_action_plan) > 0 else None
+        return (joint_action_plan[0][self.agent_index], {}) if len(joint_action_plan) > 0 else (Action.STAY, {})
 
 
 class EmbeddedPlanningAgent(Agent):
+    """
+    An agent that uses A* search to find an optimal action based on a model of the other agent,
+    `other_agent`. This class approximates the other agent as being deterministic even though it
+    might be stochastic in order to perform the search.
+    """
 
-    def __init__(self, other_agent, mlp, env, delivery_horizon=2, debug=False):
-        """other_agent_policy returns highest prob action"""
+    def __init__(self, other_agent, mlp, env, delivery_horizon=2, logging_level=0):
+        """mlp is a MediumLevelPlanner"""
         self.other_agent = other_agent
         self.delivery_horizon = delivery_horizon
         self.mlp = mlp
         self.env = env
         self.h_fn = Heuristic(mlp.mp).simple_heuristic
-        self.debug = debug
+        self.logging_level = logging_level
 
     def action(self, state):
-        from overcooked_ai_py.planning.search import SearchTree
         start_state = state.deepcopy()
-        start_state.order_list = start_state.order_list[:self.delivery_horizon]
+        order_list = start_state.order_list if start_state.order_list is not None else ["any", "any"]
+        start_state.order_list = order_list[:self.delivery_horizon]
         other_agent_index = 1 - self.agent_index
-
-        initial_other_agent_type = self.other_agent.stochastic
-        self.other_agent.stochastic = False
         initial_env_state = self.env.state
         self.other_agent.env = self.env
 
@@ -158,143 +284,222 @@ class EmbeddedPlanningAgent(Agent):
             idx = np.random.randint(5)
             return Action.ALL_ACTIONS[idx]
 
-        # Check plan
-        actions = [item[0] for item in ml_s_a_plan]
-        actions = actions[1:]
-        tru_actions = ()
-        for a in actions:
-            tru_actions = tru_actions + a
-        assert len(tru_actions) == cost
+        # Check estimated cost of the plan equals
+        # the sum of the costs of each medium-level action
+        assert sum([len(item[0]) for item in ml_s_a_plan[1:]]) == cost
 
         # In this case medium level actions are tuples of low level actions
         # We just care about the first low level action of the first med level action
         first_s_a = ml_s_a_plan[1]
 
-        # if self.debug:
-        #     print("WHAT'S ON MY MIND:")
-        #     self.env.state = start_state
-        #     for joint_a in first_s_a[0]:
-        #         print(self.env)
-        #         print(joint_a)
-        #         self.env.step(joint_a)
-        #     print(self.env)
-        #     print("======The End======")
+        # Print what the agent is expecting to happen
+        if self.logging_level >= 2:
+            self.env.state = start_state
+            for joint_a in first_s_a[0]:
+                print(self.env)
+                print(joint_a)
+                self.env.step(joint_a)
+            print(self.env)
+            print("======The End======")
 
         self.env.state = initial_env_state
-        self.other_agent.stochastic = initial_other_agent_type
 
         first_joint_action = first_s_a[0][0]
-        if self.debug: 
+        if self.logging_level >= 1:
             print("expected joint action", first_joint_action)
         action = first_joint_action[self.agent_index]
-        return action
+        return action, {}
 
-class GreedyHumanModel_mc(Agent):
+
+class GreedyHumanModel(Agent):
     """
+    This is Micah's GreedyHumanModel, which is slightly different to Paul's (called _pk)
+
     Agent that at each step selects a medium level action corresponding
     to the most intuitively high-priority thing to do
-    
-    NOTE: MIGHT NOT WORK IN ALL ENVIRONMENTS
+
+    NOTE: MIGHT NOT WORK IN ALL ENVIRONMENTS, for example forced_coordination.layout,
+    in which an individual agent cannot complete the task on their own.
     """
 
-    def __init__(self, mlp, boltzmann_rational=False, temp=1):
+    def __init__(self, mlp, hl_boltzmann_rational=False, ll_boltzmann_rational=False, hl_temp=1, ll_temp=1,
+                 auto_unstuck=True):
         self.mlp = mlp
         self.mdp = self.mlp.mdp
+
+        # Bool for perfect rationality vs Boltzmann rationality for high level and low level action selection
+        self.hl_boltzmann_rational = hl_boltzmann_rational  # For choices among high level goals of same type
+        self.ll_boltzmann_rational = ll_boltzmann_rational  # For choices about low level motion
+
+        # Coefficient for Boltzmann rationality for high level action selection
+        self.hl_temperature = hl_temp
+        self.ll_temperature = ll_temp
+
+        # Whether to automatically take an action to get the agent unstuck if it's in the same
+        # state as the previous turn. If false, the agent is history-less, while if true it has history.
+        self.auto_unstuck = auto_unstuck
+        self.reset()
+
+    def reset(self):
         self.prev_state = None
-        self.boltzmann_rational = boltzmann_rational
-        self.temperature = temp # Some measure of rationality
+
+    def actions(self, states, agent_indices):
+        actions_and_infos_n = []
+        for state, agent_idx in zip(states, agent_indices):
+            self.set_agent_index(agent_idx)
+            self.reset()
+            actions_and_infos_n.append(self.action(state))
+        return actions_and_infos_n
 
     def action(self, state):
-        motion_goals = self.ml_action(state)
+        possible_motion_goals = self.ml_action(state)
 
         # Once we have identified the motion goals for the medium
         # level action we want to perform, select the one with lowest cost
         start_pos_and_or = state.players_pos_and_or[self.agent_index]
 
-        if self.boltzmann_rational:
-            plans = [self.mlp.mp.get_plan(start_pos_and_or, goal) for goal in motion_goals]
-            action_plan = [plan[0] for plan in plans]
-            plan_cost = np.array([plan[2] for plan in plans])
-            softmax_probs = np.exp(plan_cost * self.temperature) / np.sum(np.exp(plan_cost * self.temperature))
-            goal_idx = np.random.choice(len(motion_goals), p=softmax_probs)
-            chosen_action = action_plan[goal_idx][0]
+        chosen_goal, chosen_action, action_probs = self.choose_motion_goal(start_pos_and_or, possible_motion_goals)
 
-        else:  
-            min_cost = np.Inf
-            best_action = None
-            for goal in motion_goals:
-                action_plan, _, plan_cost = self.mlp.mp.get_plan(start_pos_and_or, goal)
-                if plan_cost < min_cost:
-                    best_action = action_plan[0]
-                    min_cost = plan_cost
-            chosen_action = best_action
-        
-        # HACK: if two agents get stuck, select an action at random that would
-        # change the player positions if the other player were not to move
-        if self.prev_state is not None and state.players_pos_and_or == self.prev_state.players_pos_and_or:
-            if self.agent_index == 0:
-                joint_actions = list(itertools.product(Action.ALL_ACTIONS, [Action.STAY]))
-            elif self.agent_index == 1:
-                joint_actions = list(itertools.product([Action.STAY], Action.ALL_ACTIONS))
-            else:
-                raise ValueError("Player index not recognized")
+        if self.ll_boltzmann_rational and chosen_goal[0] == start_pos_and_or[0]:
+            chosen_action, action_probs = self.boltzmann_rational_ll_action(start_pos_and_or, chosen_goal)
 
-            unblocking_joint_actions = []
-            for j_a in joint_actions:
-                new_state, _, _ = self.mlp.mdp.get_state_transition(state, j_a)
-                if new_state.player_positions != self.prev_state.player_positions:
-                    unblocking_joint_actions.append(j_a)
+        if self.auto_unstuck:
+            # HACK: if two agents get stuck, select an action at random that would
+            # change the player positions if the other player were not to move
+            if self.prev_state is not None and state.players_pos_and_or == self.prev_state.players_pos_and_or:
+                if self.agent_index == 0:
+                    joint_actions = list(itertools.product(Action.ALL_ACTIONS, [Action.STAY]))
+                elif self.agent_index == 1:
+                    joint_actions = list(itertools.product([Action.STAY], Action.ALL_ACTIONS))
+                else:
+                    raise ValueError("Player index not recognized")
 
-            chosen_action = unblocking_joint_actions[np.random.choice(len(unblocking_joint_actions))][self.agent_index]
+                unblocking_joint_actions = []
+                for j_a in joint_actions:
+                    new_state, _, _ = self.mlp.mdp.get_state_transition(state, j_a)
+                    if new_state.player_positions != self.prev_state.player_positions:
+                        unblocking_joint_actions.append(j_a)
 
-        # NOTE: Assumes that calls to action are sequential
-        self.prev_state = state
-        return chosen_action
+                chosen_action = unblocking_joint_actions[np.random.choice(len(unblocking_joint_actions))][
+                    self.agent_index]
+                action_probs = self.a_probs_from_action(chosen_action)
+
+            # NOTE: Assumes that calls to the action method are sequential
+            self.prev_state = state
+        return chosen_action, {"action_probs": action_probs}
+
+    def choose_motion_goal(self, start_pos_and_or, motion_goals):
+        """
+        For each motion goal, consider the optimal motion plan that reaches the desired location.
+        Based on the plan's cost, the method chooses a motion goal (either boltzmann rationally
+        or rationally), and returns the plan and the corresponding first action on that plan.
+        """
+        if self.hl_boltzmann_rational:
+            possible_plans = [self.mlp.mp.get_plan(start_pos_and_or, goal) for goal in motion_goals]
+            plan_costs = [plan[2] for plan in possible_plans]
+            goal_idx, action_probs = self.get_boltzmann_rational_action_idx(plan_costs, self.hl_temperature)
+            chosen_goal = motion_goals[goal_idx]
+            chosen_goal_action = possible_plans[goal_idx][0][0]
+        else:
+            chosen_goal, chosen_goal_action = self.get_lowest_cost_action_and_goal(start_pos_and_or, motion_goals)
+            action_probs = self.a_probs_from_action(chosen_goal_action)
+        return chosen_goal, chosen_goal_action, action_probs
+
+    def get_boltzmann_rational_action_idx(self, costs, temperature):
+        """Chooses index based on softmax probabilities obtained from cost array"""
+        costs = np.array(costs)
+        softmax_probs = np.exp(-costs * temperature) / np.sum(np.exp(-costs * temperature))
+        action_idx = np.random.choice(len(costs), p=softmax_probs)
+        return action_idx, softmax_probs
+
+    def get_lowest_cost_action_and_goal(self, start_pos_and_or, motion_goals):
+        """
+        Chooses motion goal that has the lowest cost action plan.
+        Returns the motion goal itself and the first action on the plan.
+        """
+        min_cost = np.Inf
+        best_action, best_goal = None, None
+        for goal in motion_goals:
+            action_plan, _, plan_cost = self.mlp.mp.get_plan(start_pos_and_or, goal)
+            if plan_cost < min_cost:
+                best_action = action_plan[0]
+                min_cost = plan_cost
+                best_goal = goal
+        return best_goal, best_action
+
+    def boltzmann_rational_ll_action(self, start_pos_and_or, goal, inverted_costs=False):
+        """
+        Computes the plan cost to reach the goal after taking each possible low level action.
+        Selects a low level action boltzmann rationally based on the one-step-ahead plan costs.
+        If `inverted_costs` is True, it will make a boltzmann "irrational" choice, exponentially
+        favouring high cost plans rather than low cost ones.
+        """
+        future_costs = []
+        for action in Action.ALL_ACTIONS:
+            pos, orient = start_pos_and_or
+            new_pos_and_or = self.mdp._move_if_direction(pos, orient, action)
+            _, _, plan_cost = self.mlp.mp.get_plan(new_pos_and_or, goal)
+            sign = (-1) ** int(inverted_costs)
+            future_costs.append(sign * plan_cost)
+
+        action_idx, action_probs = self.get_boltzmann_rational_action_idx(future_costs, self.ll_temperature)
+        return Action.ALL_ACTIONS[action_idx], action_probs
 
     def ml_action(self, state):
-        """Selects a medium level action for the current state"""
+        """
+        Selects a medium level action for the current state.
+        Motion goals can be thought of instructions of the form:
+            [do X] at location [Y]
+
+        In this method, X (e.g. deliver the soup, pick up an onion, etc) is chosen based on
+        a simple set of greedy heuristics based on the current state.
+
+        Effectively, will return a list of all possible locations Y in which the selected
+        medium level action X can be performed.
+        """
         player = state.players[self.agent_index]
         other_player = state.players[1 - self.agent_index]
         am = self.mlp.ml_action_manager
-        
+
         counter_objects = self.mlp.mdp.get_counter_objects_dict(state, list(self.mlp.mdp.terrain_pos_dict['X']))
         pot_states_dict = self.mlp.mdp.get_pot_states(state)
 
-        # TODO: this most likely will fail in some tomato scenarios
-        next_order = state.order_list[0]
+        # NOTE: this most likely will fail in some tomato scenarios
+        curr_order = state.curr_order
 
         if not player.has_object():
 
-            if next_order == 'any':
+            if curr_order == 'any':
                 ready_soups = pot_states_dict['onion']['ready'] + pot_states_dict['tomato']['ready']
                 cooking_soups = pot_states_dict['onion']['cooking'] + pot_states_dict['tomato']['cooking']
             else:
-                ready_soups = pot_states_dict[next_order]['ready']
-                cooking_soups = pot_states_dict[next_order]['cooking']
-            
+                ready_soups = pot_states_dict[curr_order]['ready']
+                cooking_soups = pot_states_dict[curr_order]['cooking']
+
             soup_nearly_ready = len(ready_soups) > 0 or len(cooking_soups) > 0
             other_has_dish = other_player.has_object() and other_player.get_object().name == 'dish'
-            
+
             if soup_nearly_ready and not other_has_dish:
                 motion_goals = am.pickup_dish_actions(state, counter_objects)
             else:
                 next_order = None
-                if len(state.order_list) > 1:
-                    next_order = state.order_list[1]
-                
+                if state.num_orders_remaining > 1:
+                    next_order = state.next_order
+
                 if next_order == 'onion':
                     motion_goals = am.pickup_onion_actions(state, counter_objects)
                 elif next_order == 'tomato':
                     motion_goals = am.pickup_tomato_actions(state, counter_objects)
                 elif next_order is None or next_order == 'any':
-                    motion_goals = am.pickup_onion_actions(state, counter_objects) + am.pickup_tomato_actions(state, counter_objects)
+                    motion_goals = am.pickup_onion_actions(state, counter_objects) + am.pickup_tomato_actions(state,
+                                                                                                              counter_objects)
 
         else:
             player_obj = player.get_object()
-            
+
             if player_obj.name == 'onion':
                 motion_goals = am.put_onion_in_pot_actions(pot_states_dict)
-            
+
             elif player_obj.name == 'tomato':
                 motion_goals = am.put_tomato_in_pot_actions(pot_states_dict)
 
@@ -306,12 +511,13 @@ class GreedyHumanModel_mc(Agent):
 
             else:
                 raise ValueError()
-        
-        motion_goals = list(filter(lambda goal: self.mlp.mp.is_valid_motion_start_goal_pair(player.pos_and_or, goal), motion_goals))
+
+        motion_goals = [mg for mg in motion_goals if self.mlp.mp.is_valid_motion_start_goal_pair(player.pos_and_or, mg)]
 
         if len(motion_goals) == 0:
             motion_goals = am.go_to_closest_feature_actions(player)
-            motion_goals = list(filter(lambda goal: self.mlp.mp.is_valid_motion_start_goal_pair(player.pos_and_or, goal), motion_goals))
+            motion_goals = [mg for mg in motion_goals if
+                            self.mlp.mp.is_valid_motion_start_goal_pair(player.pos_and_or, mg)]
             assert len(motion_goals) != 0
 
         return motion_goals
@@ -356,7 +562,7 @@ class AgentPair(object):
             joint_action_plan = self.a0.mlp.get_low_level_action_plan(state, self.a0.heuristic, delivery_horizon=self.a0.delivery_horizon, goal_info=True)
             return joint_action_plan[0] if len(joint_action_plan) > 0 else (None, None)
         elif self.a0 is self.a1:
-            # When using the same instance of an agent for self-play, 
+            # When using the same instance of an agent for self-play,
             # reset agent index at each turn to prevent overwriting it
             self.a0.set_agent_index(0)
             action_0 = self.a0.action(state)
@@ -365,7 +571,7 @@ class AgentPair(object):
             return (action_0, action_1)
         else:
             return (self.a0.action(state), self.a1.action(state))
-        
+
     def reset(self):
         for a in self.agents:
             a.reset()
@@ -373,8 +579,10 @@ class AgentPair(object):
 
 # ============================ Agents make by pk ===================================#
 
-class GreedyHumanModelv2(Agent):
+class GreedyHumanModel_pk(Agent):
     """
+    This is Paul's GreedyHumanModel, which is slightly different Micah's (just called GreedyHumanModel)
+
     Agent that at each step selects a medium level action corresponding to the most intuitively high-priority thing to do
 
     Enhancements added to v2:
@@ -611,7 +819,7 @@ class ToMModel(Agent):
         self.only_take_dispenser_onions = False  # Only used for layout Forced Coordination: when an onion is
         # deliberately dropped on a shared counter, then from then on only take onions from the dispenser
         self.only_take_dispenser_dishes = False
-        self.GHM = GreedyHumanModelv2(self.mlp, player_index=1-self.agent_index)  # For ToM of other players
+        self.GHM = GreedyHumanModel_pk(self.mlp, player_index=1 - self.agent_index)  # For ToM of other players
         self.human_model = True
 
         # "Personality" parameters (within 0 to 1, except rationality_coefficient)
